Fix a compilation error (ambiguous cast from Double to ?). This simply changes the code to directly use uint as that's all that is needed.

+++ lib/wormux/tools/random.cpp
@@ -93,8 +93,8 @@ uint RandomGenerator::GetRand()
 
 bool RandomGenerator::GetBool()
 {
-  Double middle = WORMUX_RAND_MAX/2;
-  Double random = GetRand();
+  uint middle = WORMUX_RAND_MAX/2;
+  uint random = GetRand();
   return (random <= middle);
 }
